+++ src/main/java/de/cismet/cids/custom/wunda_blau/band/TreppeBandMember.java
@@ -78,6 +78,7 @@ public abstract class TreppeBandMember extends JXPanel implements ModifiableBand
     protected List<ElementResizedListener> elementResizeListener = new ArrayList<ElementResizedListener>();
     protected boolean dragStart = false;
     protected TreppenBand parent;
+    protected boolean alternativeColor = false;
 
     double von = 0;
     double bis = 0;
@@ -161,6 +162,16 @@ public abstract class TreppeBandMember extends JXPanel implements ModifiableBand
     /**
      * DOCUMENT ME!
      *
+     * @param  alternativeColor  DOCUMENT ME!
+     */
+    public void setAlternativeColor(final boolean alternativeColor) {
+        this.alternativeColor = alternativeColor;
+        determineBackgroundColour();
+    }
+
+    /**
+     * DOCUMENT ME!
+     *
      * @param  text  DOCUMENT ME!
      */
     public void setText(final String text) {
@@ -220,34 +231,61 @@ public abstract class TreppeBandMember extends JXPanel implements ModifiableBand
     /**
      * DOCUMENT ME!
      */
+    protected void setReadOnlyColor() {
+        unselectedBackgroundPainter = new MattePainter(new Color(75, 75, 75));
+        selectedBackgroundPainter = new CompoundPainter(
+                unselectedBackgroundPainter,
+                new RectanglePainter(
+                    3,
+                    3,
+                    3,
+                    3,
+                    3,
+                    3,
+                    true,
+                    new Color(100, 100, 100, 100),
+                    2f,
+                    new Color(50, 50, 50, 100)));
+        if (isSelected) {
+            setBackgroundPainter(selectedBackgroundPainter);
+        } else {
+            setBackgroundPainter(unselectedBackgroundPainter);
+        }
+    }
+
+    /**
+     * DOCUMENT ME!
+     */
     protected void configurePopupMenu() {
         popup.removeAll();
-        final JMenuItem splitItem = new JMenuItem();
-        splitItem.setAction(new SplitItem(this));
-        final JMenuItem deleteItem = new JMenuItem();
-        deleteItem.setAction(new DeleteItem(this));
+        if (!isReadOnly()) {
+            final JMenuItem splitItem = new JMenuItem();
+            splitItem.setAction(new SplitItem(this));
+            final JMenuItem deleteItem = new JMenuItem();
+            deleteItem.setAction(new DeleteItem(this));
 
-        final String[] objectsNames = parent.getAllowedObjectNames();
-        final String[] objectsTables = parent.getAllowedObjectTableNames();
+            final String[] objectsNames = parent.getAllowedObjectNames();
+            final String[] objectsTables = parent.getAllowedObjectTableNames();
 
-        JMenuItem item = new JMenuItem();
-        item.setAction(new AddItem(this, false, objectsTables[0], objectsNames[0] + " davor hinzufügen"));
-        if ((getMin() == 0.0) || (parent.getNextLessElement(this) != null)) {
-            item.setEnabled(false);
+            JMenuItem item = new JMenuItem();
+            item.setAction(new AddItem(this, false, objectsTables[0], objectsNames[0] + " davor hinzufügen"));
+            if ((getMin() == 0.0) || (parent.getNextLessElement(this) != null)) {
+                item.setEnabled(false);
+            }
+            popup.add(item);
+
+            item = new JMenuItem();
+            item.setAction(new AddItem(this, true, objectsTables[0], objectsNames[0] + " danach hinzufügen"));
+            if (parent.getNextGreaterElement(this) != null) {
+                item.setEnabled(false);
+            }
+            popup.add(item);
+
+            popup.addSeparator();
+            popup.add(splitItem);
+            popup.addSeparator();
+            popup.add(deleteItem);
         }
-        popup.add(item);
-
-        item = new JMenuItem();
-        item.setAction(new AddItem(this, true, objectsTables[0], objectsNames[0] + " danach hinzufügen"));
-        if (parent.getNextGreaterElement(this) != null) {
-            item.setEnabled(false);
-        }
-        popup.add(item);
-
-        popup.addSeparator();
-        popup.add(splitItem);
-        popup.addSeparator();
-        popup.add(deleteItem);
     }
 
     /**
